Move DAS code to das subdirectory
