+++ DAS/das/tests/__init__.py
@@ -1,0 +1,3 @@
+from .das_data_generator import send_fake_data, send_csv_data
+
+__all__ = ["send_fake_data", "send_csv_data"]
