+++ DAS/pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "DAS"
+name = "das"
 version = "0.1.0"
 description = ""
 authors = ["Blake <bhay0010@student.monash.edu>"]
