--- DAS/tests/__init__.py
@@ -1,1 +1,0 @@
-from DAS.utils.MockSensor import MockSensor
